Move join() to korolev.effect.Queue #348
